+++ Runtime/Scripts/Simple/FPEVManager.cs
@@ -23,11 +23,30 @@ namespace FuzzPhyte.SGraph
         public HelperCategory HelperType;
         
         public float TimeUntil;
+        //Unique Tag
+        public FP_Data TargetObjectData;
+        public FPEventActionType ActionType;
+        [Tooltip("Used for state change on enabled/disabled items")]
+        public bool BoolActionTypeState;
+        [Tooltip("Used for method name and/or other lookup references for a component type")]
+        public string CustomString_NameAction;
         public UnityEvent TheHelperAction;
         public void ActivateAction()
         {
             TheHelperAction.Invoke();
         }
+    }
+    [Serializable]
+    public enum FPEventActionType
+    {
+        //Do Nothing
+        NA=0,
+        //SetActive(true)/SetActive(false)
+        SetActive = 1,
+        //enabled = true / enabled = false
+        ComponentActive = 2,
+        PlayAnimationTrigger = 8,
+        CustomMethod = 9,
     }
     #endregion
     public class FPEVManager : MonoBehaviour
@@ -39,6 +58,7 @@ namespace FuzzPhyte.SGraph
             eventStates = new Dictionary<FPMonoEvent, FPEventState>();
         }
         #region Standard Event Functions
+        
         public virtual void AddFPEventStateData(FPMonoEvent theKey,FPEventState eventState)
         {
             if(eventStates.ContainsKey(theKey))
@@ -60,6 +80,21 @@ namespace FuzzPhyte.SGraph
                 Debug.LogWarning($"Key,{theKey.name} doesn't exist in the dictionary");
             }
         }
+        /// <summary>
+        /// Method for us to invoke after we have setup our event and we need to fire off the immediate Unity Action associated with it
+        /// </summary>
+        /// <param name="theEventKey"></param>
+        public virtual void TriggerEventStateSetup(FPMonoEvent theEventKey)
+        {
+            if (!eventStates.ContainsKey(theEventKey))
+            {
+                Debug.LogWarning("Key not found in the dictionary");
+                return;
+            }
+            var initTrue = eventStates[theEventKey].TryInvokeEventInitialization();
+            var curState = eventStates[theEventKey].CurrentState;
+            theEventKey.PassBackFromManager(initTrue, curState);
+        }
         public virtual void TriggerEventTransition(FPMonoEvent theEventKey, SequenceTransition transition, List<RequirementD> requirementValue)
         {
             if(!eventStates.ContainsKey(theEventKey))
@@ -80,6 +115,7 @@ namespace FuzzPhyte.SGraph
             var returnValues = eventStates[theEventKey].TryTransition(transition);
             theEventKey.PassBackFromManager(returnValues.Item1,returnValues.Item2);
         }
+        
         #endregion
     }
 }